feat(design/template): restructure file system
-rename db folder models
-rename helpers folder middelwares
-create a route folder to hold all routes

[delivers #163577871]
