+++ source/app.js
@@ -1,6 +1,7 @@
 import express from 'express';
 import bodyParser from 'body-parser';
-import route from './routes';
+import parties from './routes/parties';
+import offices from './routes/offices';
 
 const PORT = process.env.PORT || 5000;
 
@@ -9,7 +10,8 @@ const app = express();
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
-app.use('/api/v1', route);
+app.use('/api/v1/parties', parties);
+app.use('/api/v1/offices', offices);
 app.use((req, res, next) => {
   const error = new Error('Not found');
   error.status = 404;
